passed through moulinette

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: bde-koni <bde-koni@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/08 17:01:22 by bde-koni          #+#    #+#             */
-/*   Updated: 2024/11/05 18:39:09 by bde-koni         ###   ########.fr       */
+/*   Updated: 2024/11/19 15:24:38 by bde-koni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdio.h>
+#include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
@@ -27,7 +27,7 @@ int	ft_strlen(const char *str)
 // int main(void)
 // {
 //     char str[55] = "vier";
-
+//
 //     printf("%d \n", ft_strlen(str));
 //     return (0);
 // }
